requirements.txt
# Contains all libraries that are needed for this project

# search api
tavily-python
load_dotenv
amadeus

#For Langraph 
graphviz
pygraphviz

# LLM apis
openai

# to load env files
python-dotenv

# langchain libraries
langchain
langchain_community
wikipedia